move email form test to contact test file
